feat(suite): add buy/sell and swap buttons to dashboard header
